requirements.txt
fastapi==0.55.1
fastapi-sqlalchemy
pydantic
alembic
psycopg2
uvicorn==0.11.5
python-dotenv
jinja2
python-multipart
python-socketio
redis
httpx
pytest
requests